Reducing the annual salary
I originally chose £180k a year because it is a huge amount, but not outside the realms of posibility.

Almost everybody who gave me feedback said this was too much and made the valid point that no prospective MP would publically sign up to a 100% pay rise.

Thus, I've reduced it the a salary that's roughly what I see advertised for Director posts in IT companies, people who often have tens of staff and budgetary responsibilities

Quotes verbatim:

"The salary is far too high, I couldn’t justify it."

"In relation to MP’s pay, no issue with linking it to the minimum wage or average wages, but no MP/candidate is going to sign a pledge which calls for a near 100% pay rise. This pledge could be changed to link the minimum wage to MPs’ current salaries, so that in future, MP pay has a link to the wealth and success of the economy and how it rewards the lowest-paid workers."